Drop dual PHP/PHP7 executors as per #117
