--- dmoj/executors/PHP7.py
@@ -1,7 +1,0 @@
-from .php_executor import PHPExecutor
-
-
-class Executor(PHPExecutor):
-    name = 'PHP7'
-    command = 'php7'
-    fs = ['.*\.ini$']
